refactor(auth): consolidate five auth modules into core/auth.py
authentication.py, oauth2_validators.py, oidc_verify.py and tokens.py were
one class or one object each, and finding anything auth-related meant opening
four files. Merge them into a single core/auth.py with numbered sections
(activation tokens, DRF authentication, DOT OIDC validator, advertised OAuth
endpoint URIs, EHR id_token verification). None of the old filenames was
required by Django or DRF -- both settings references are dotted paths, now
repointed at core.auth.

core/models/jhe_user.py must import the activation token lazily: core.auth
imports DOT's oauth2_validators, which calls get_application_model() at module
scope, and core is first in INSTALLED_APPS -- so a top-level import would run
before the app registry is ready (AppRegistryNotReady).
