--- core/tokens.py
@@ -1,10 +1,0 @@
-from django.contrib.auth.tokens import PasswordResetTokenGenerator
-from six import text_type
-
-
-class TokenGenerator(PasswordResetTokenGenerator):
-    def _make_hash_value(self, user, timestamp):
-        return text_type(user.id) + text_type(timestamp) + text_type(user.email_is_verified)
-
-
-account_activation_token = TokenGenerator()
