+++ tests/backend/test_token_exchange.py
@@ -5,7 +5,7 @@ import pytest
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
 
-from core import oidc_verify
+from core import auth
 
 ISS = "https://ehr.example.org/fhir"
 AUD = "smart-client-id"
@@ -80,8 +80,8 @@ def patch_jwks(monkeypatch, rsa_private_pem):
         def get_signing_key_from_jwt(self, token):
             return _SigningKey()
 
-    monkeypatch.setattr(oidc_verify, "discover_jwks_uri", lambda issuer: "https://ehr.example.org/jwks")
-    monkeypatch.setattr(oidc_verify, "_jwk_client", lambda jwks_uri: _FakeClient())
+    monkeypatch.setattr(auth, "discover_jwks_uri", lambda issuer: "https://ehr.example.org/jwks")
+    monkeypatch.setattr(auth, "_jwk_client", lambda jwks_uri: _FakeClient())
 
 
 def make_token(
